Add new GameState: `GameStartsSoon`
This new state is between the transition from
`AwaitingPlayers` and `InProgress`. The purpose of this new state
is to let players be ready: when transiting to this state, the server
will send a count down duration on when the game will start.

Problems to be fixed:
- After a player cleaned his board, it does not broadcast to everyone
and does not disconnect others from websockets
- When a player is disconnect from a websocket, in the server side, it
throws kind of exception, i dont know if it should happen or not.